Test the bits, not the number, in the instruction writer
Three wrong answers from a sweep aimed at the operators, all the
instruction writer's, and one of them in every fixture with a discard.

`if (r0.y != 0)` for if_nz and `(r0.y != 0) ?` for movc test the
float, and the instruction tests the bits. r0.y held a sign bit,
`and r0.y, r0.x, l(0x80000000)`, which as a float is -0.0, and
-0.0 != 0 is false. Every test against zero is `asint(x) != 0` now
wherever the register is not declared int, and the constant's own type
where it is a constant - `useDetail != 0` for a bool, not
`asint(useDetail)`; an immediate is its bits.

`clip(r0.x)` for discard_nz tests the sign of a float, and a comparison
mask is NaN as a float, which is never negative: the writer's output
never discarded anything, in clip_alpha, clip_loop and conditional_depth
alike. The equivalence tier had not noticed, because a discarded pixel
has no outputs and it compared the outputs both sides had - none. A
discard is an outcome now, one side discarding and the other not a
difference, and the three failed the moment it was and pass with
`if (asint(x) != 0) discard;`.

Then udiv. `asint(x) % 5` is a signed modulus, which fxc lowers with a
sign test of its own and merges with the one the shader already had;
the operands are unsigned now, `asuint(x) % 5`, as wide as the result.
And `asfloat(u % 5)` - an unsigned remainder by an immediate,
reinterpreted - fxc folds to zero outright, a bug of its own; spelled
`u - u / 5 * 5` it compiles, so that is how a reinterpreted remainder is
written. `ine` was unimplemented in this writer alone, one case away
from ieq.

+++ HlslDecompiler.Tests/ShaderSources/ps_4_0_instruction/int_constant_divide.fx
@@ -10,8 +10,8 @@ float4 main(float4 texcoord : TEXCOORD) : SV_Target
 	r0.x = value ^ 7;
 	r0.x = r0.x & -2147483648;
 	r0.y = max(value, -(value));
-	r0.z = r0.y / 7;
-	r0.y = r0.y % 13;
+	r0.z = (uint)r0.y / 7;
+	r0.y = (uint)r0.y % 13;
 	r0.w = -r0.z;
 	r0.x = (r0.x != 0) ? r0.w : r0.z;
 	r1.x = (float)(int)r0.x;
@@ -19,9 +19,9 @@ float4 main(float4 texcoord : TEXCOORD) : SV_Target
 	r0.z = value & -2147483648;
 	r0.x = (r0.z != 0) ? r0.x : r0.y;
 	r1.y = (float)(int)r0.x;
-	r0.x = uvalue / 10;
+	r0.x = (uint)uvalue / 10;
 	r1.z = (float)(uint)r0.x;
-	r0.x = uvalue % 6;
+	r0.x = (uint)uvalue % 6;
 	r1.w = (float)(uint)r0.x;
 	o = r1 * texcoord;
